fix: simplify WebDeployment custom resource building
