+++ packages/gboost-infra/scripts/build.ts
@@ -1,20 +1,22 @@
-import { execSync } from "node:child_process";
 import { fileURLToPath } from "node:url";
 import { cpSync, rmSync, statSync } from "node:fs";
 import { resolve } from "node:path";
-import { buildWebDeploymentCr } from "./build-web-deployment-cr.js";
+import { buildFunction } from "./build-function.js";
+import { execSync } from "node:child_process";
 
 const thisFilePath = fileURLToPath(import.meta.url);
-
 const libDir = resolve(thisFilePath, "../../lib");
 const srcDir = resolve(thisFilePath, "../../src");
+
 rmSync(libDir, { recursive: true, force: true });
 execSync("pnpm exec tsc -p tsconfig.build.json", { stdio: "inherit" });
-// bundle and minify WebDeployment custom resource lambda with esbuild
-// (./src/web-deployment/custom-resource-handler)
+// bundle and minify functions with esbuild
 // Note, tsconfig.build.json ignores these files so tsc transpiled js files
 // don't overrwrite esbuild built files
-await buildWebDeploymentCr();
+await buildFunction({
+  entryPointPath: "web-deployment/web-deploy-cr-handler.ts",
+  outFilePath: "web-deployment/web-deploy-cr-handler.mjs",
+});
 // needed for vtl and js files that aren't built with tsc
 cpSync(srcDir, libDir, {
   filter(source: string, destination: string) {
