+++ packages/gboost-infra/src/web-deployment/web-deployment.ts
@@ -144,7 +144,7 @@ export class WebDeployment extends Construct {
       prune: props.prune || true,
     };
 
-    this.#getCustomResource({ fn, properties });
+    this.#createCustomResource({ fn, properties });
     this.#suppressNags(fn);
   }
 
@@ -177,10 +177,10 @@ export class WebDeployment extends Construct {
     if (isDev) {
       codePath = resolve(
         thisFilePath,
-        "../../../lib/web-deployment/custom-resource-handler"
+        "../../../lib/web-deployment/web-deploy-cr-handler"
       );
     } else {
-      codePath = resolve(thisFilePath, "../custom-resource-handler");
+      codePath = resolve(thisFilePath, "../web-deploy-cr-handler-handler");
     }
     const uuid = "e94797a6-488d-4a73-8e2b-79695c7ec7dd";
     const lambdaPurpose = "WebDeploymentCustomResourceHandler";
@@ -236,7 +236,7 @@ export class WebDeployment extends Construct {
     });
   }
 
-  #getCustomResource(params: GetCustomResourceParams) {
+  #createCustomResource(params: GetCustomResourceParams) {
     const provider = new Provider(this, "Provider", {
       onEventHandler: params.fn,
       logRetention: RetentionDays.ONE_DAY,
