+++ packages/gboost-infra/tsconfig.build.json
@@ -1,5 +1,9 @@
 {
   "extends": "./tsconfig.json",
   "include": ["src/**/*"], // don't compile scripts/
-  "exclude": ["**/*.test.ts", "lib", "./src/web-deployment/custom-resource-handler/**"]
+  "exclude": [
+    "**/*.test.ts",
+    "lib",
+    "./src/web-deployment/web-deploy-cr-handler/**",
+  ]
 }
